feat: move reason for removal transformations into separate class
