+++ application/CohortManager/src/Functions/CohortDistributionServices/TransformDataService/ITransformReasonForRemoval.cs
@@ -1,0 +1,7 @@
+namespace NHS.CohortManager.CohortDistribution;
+
+using Model;
+public interface ITransformReasonForRemoval
+{
+    Task<CohortDistributionParticipant> ReasonForRemovalTransformations(CohortDistributionParticipant participant);
+}
